+++ src/SampleTest/Linq/EnumerableTest.cs
@@ -46,30 +46,6 @@ public class EnumerableTest {
 	}
 
 	[Fact]
-	public void Prepend_シーケンスの最初に要素を追加する() {
-		// Arrange
-		var source = new[] { 2, 3, 4 };
-
-		// Act
-		var actual = source.Prepend(1);
-
-		// Assert
-		Assert.Equal([1, 2, 3, 4], actual);
-	}
-
-	[Fact]
-	public void Append_シーケンスの最後に要素を追加する() {
-		// Arrange
-		var source = new[] { 2, 3, 4 };
-
-		// Act
-		var actual = source.Append(5);
-
-		// Assert
-		Assert.Equal([2, 3, 4, 5], actual);
-	}
-
-	[Fact]
 	public void Min_非nullの空のシーケンスで呼び出すとInvalidOperationException() {
 		// 非nullの場合は例外
 		Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<int>().Min());
